Build on 64 bit MacOS with -Wall and -Werror. make install failing trying to call adafront to build predef - "chaos: emalloc: ridiculous argument"

+++ c/adaed.1.11.2/nyu/Adaed-1.11.2/6.c
@@ -220,7 +220,7 @@ Tuple get_formals(Node formal_list, char *proc_id) 			/*;get_formals*/
 			formal_index++;
 			tup = tup_new(4);
 			tup[1] = (char *)N_VAL(id_node);
-			tup[2] = (char *) f_mode;
+			tup[2] = (char *)(long long) f_mode;
 			tup[3] = (char *) type_mark;
 			tup[4] = (char *) copy_tree(exp_node);
 			formals[formal_index] = (char *) tup;
